# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ skeleSim/R/runSim.R
@@ -4,6 +4,7 @@
 #' @param params a \linkS4class{skeleSim.params} object.
 #' @param num.secs number of seconds to run timing checks
 #'
+#' @importFrom swfscMisc autoUnits
 #' @export
 #'
 runSim <- function(params, num.secs = NULL) {
@@ -21,9 +22,7 @@ runSim <- function(params, num.secs = NULL) {
 print("params checked")
 
   # Check/setup folder structure
-  if(file.exists(params@wd)) {
-    unlink(params@wd, recursive = TRUE, force = TRUE)
-  }
+  if(file.exists(params@wd)) unlink(params@wd, recursive = TRUE, force = TRUE)
   dir.create(params@wd)
   wd <- setwd(params@wd)
 
@@ -32,9 +31,7 @@ print("params checked")
   tryCatch({
     num.sc <- length(params@scenarios)
     num.reps <- params@num.reps
-    sc.vec <- rep(1:num.sc, num.reps)
-    rep.vec <- rep(1:num.reps, each = num.sc)
-    params@scenario.reps <- cbind(scenario = sc.vec, replicate = rep.vec)
+    params@scenario.reps <- as.matrix(expand.grid(scenario = 1:num.sc, replicate = 1:num.reps))
     quit <- FALSE
     # loop through replicates for scenarios
     num.iter <- nrow(params@scenario.reps)
@@ -52,6 +49,7 @@ print("params checked")
       file <- paste(label, ".params.rdata", sep = "")
       if(!dir.exists(label)) dir.create(label)
       save(params, file = file.path(label, file))
+#-----
       # check timing
       results$timing$end.time <- Sys.time()
       if(!is.null(num.secs)) {
@@ -62,21 +60,7 @@ print("params checked")
     }
 
     elapsed <- results$timing$end.time - results$timing$start.time
-    units(elapsed) <- "secs"
-    completion <- num.iter * elapsed / i
-    if(completion > 60) {
-      units(completion) <- "mins"
-      if(completion > 60) {
-        units(completion) <- "hours"
-        if(completion > 24) {
-          units(completion) <- "days"
-          if(completion > 7) {
-            units(completion) <- "weeks"
-          }
-        }
-      }
-    }
-    results$timing$completion.time <- completion
+    results$timing$completion.time <- autoUnits(num.iter * elapsed / i)
     results$timing$pct.complete <- round(100 * i / num.iter, 1)
     results$params <- params
   }, finally = setwd(wd))
